Proper suppoort for SQL keywords (closes #2)
Fastparse doesn't have first class support for keywords.  Previously,
they were hacked in using StringInIgnoreCase, with a specific list of
keywords that couldn't be taken as identifiers.  #2 however shows the
opposite case: identifiers being taken as keywords.

I've finally figured out a reasonably clean way to capture keywords:
There's now Elements.anyKeyword, which captures every single keyword
used by Sparsity in a StringIsIgnoreCase, followed by a rule that
rejects the keyword if it is immediately followed by a keyword-
compatible character ( /[a-zA-Z0-9_]/ ).

All instances of StringInIgnoreCase have now been replaced by
Elements.keyword (aliased to uppercase Keyword in both the SQL and
Expression parsers), which takes a list of acceptable keywords. The
function uses anyKeyword to read a single keyword token, and manually
checks to see if the token matches one of the arguments.  It's a
little ugly/inefficient, but it's the closest I can come to something
reasonable without hacking FastParse itself.